Organized minor components better, removed lots of unnecessary scaffolding
