+++ app/assets/javascripts/megatron/utils/clipboard.js
@@ -28,4 +28,4 @@ var Clipboard = {
   }
 }
 
-module.exports = Clipboard
+Event.change(Clipboard.setup)
