+++ app/assets/javascripts/megatron/utils/form-notify.js
@@ -1,39 +1,33 @@
-var event = require('compose-event')
-var notify = require('notify')
+var Event = require('compose-event')
+var Notify = require('notify')
 
-module.exports = {
-  notify: function formNotify(el, type){
-    if(type == 'success')
-      notify.success(this.getMessage(el, 'success'))
-    else if(type == 'error')
-      notify.error(this.getMessage(el, 'error'))
-    else if(type == 'beforeSend')
-      notify.progress(this.getMessage(el, 'beforeSend'))
-  },
-
-  getMessage: function formGetMessage(el, type){
-    if (el.dataset[type])
-      return el.dataset[type]
-    else {
-      var msg_object = document.querySelector('script.'+type)
-
-      if (msg_object) {
-        var msg = msg_object.innerHTML
-        if (msg)
-          return msg
-        else
-          return this.defaultMessages[type]
-      }
-    }
-  },
-
-  listen: function formListen() {
-    event.on(document.querySelector('body'), 'submit', 'form', function(event){
-      if(!event.target.dataset['remote']) {
-        event.preventDefault()
-        setTimeout(function(){event.target.submit()}, 100)
-        this.notify(event.target, 'beforeSend')
-      }
-    }.bind(this))
-  }
+var defaultMessages = {
+  beforeSend: 'Submitting...',
+  success: 'Success!',
+  error: 'Something went wrong.'
 }
+
+var notifyForm = function(event) {
+  var message = getMessage(event.target, event.type)
+
+  if (event.type == 'beforeSend')
+    Notify.progress(message)
+  else
+    Notify[type](message)
+}
+
+var getMessage = function(form, type) {
+  if (form.dataset[type])
+    return form.dataset[type]
+
+  var el = form.querySelector('[data-ajax-event='+type+'], script.'+type)
+
+  if (el)
+    return el.innerHTML
+  else
+    return defaultMessages[type]
+}
+
+Event.ready(function() {
+  Event.on(document, 'beforeSend success error', 'form[data-remote]', notifyForm)
+})
