+++ package.json
@@ -19,7 +19,7 @@
     "codemirror": "^5.5.0",
     "compose-dialog": "^2.0.0",
     "compose-notification": "^3.0.0",
-    "compose-remote-form": "^1.0.2",
+    "compose-remote-form": "^2.0.0",
     "compose-time-toggle": "^1.0.0",
     "compose-toggler": "^1.0.0",
     "domify": "^1.3.3",
